requirements.txt
fastapi[all]
uvicorn
sqlalchemy
alembic
gunicorn
python-dotenv
psycopg2
psycopg2-binary